inspector/requirements.txt
Flask
ffmpeg-python
requests 